:white_check_mark: add test launch protocol from singed out state
