+++ test/cypress/utils/loginUI.js
@@ -1,0 +1,10 @@
+import ss from '../../constants/selectors';
+
+export default function loginUI(passphrase) {
+  cy.get(ss.passphraseInput).first().click();
+  cy.get(ss.passphraseInput).each(($el, index) => {
+    const passphraseWordsArray = passphrase.split(' ');
+    cy.wrap($el).type(passphraseWordsArray[index]);
+  });
+  cy.get(ss.loginBtn).click();
+}
